+++ bin/prettier.js
@@ -98,7 +98,8 @@ function getParserOption() {
     value === "flow" ||
     value === "babylon" ||
     value === "typescript" ||
-    value === "postcss"
+    value === "postcss" ||
+    value === "graphql"
   ) {
     return value;
   }
@@ -203,7 +204,7 @@ function format(input, opt) {
           diff(input, pp);
       }
     }
-    return {};
+    return { formatted: opt.filepath || "(stdin)\n" };
   }
 
   return prettier.formatWithCursor(input, opt);
@@ -282,7 +283,7 @@ if (stdin) {
   });
 } else {
   eachFilename(filepatterns, filename => {
-    if (write || argv["debug-check"]) {
+    if (write) {
       // Don't use `console.log` here since we need to replace this line.
       process.stdout.write(filename);
     }
@@ -301,7 +302,12 @@ if (stdin) {
     }
 
     if (argv["list-different"]) {
-      if (!prettier.check(input, options)) {
+      if (
+        !prettier.check(
+          input,
+          Object.assign({}, options, { filepath: filename })
+        )
+      ) {
         if (!write) {
           console.log(filename);
         }
@@ -355,7 +361,6 @@ if (stdin) {
         }
       }
     } else if (argv["debug-check"]) {
-      process.stdout.write("\n");
       if (output) {
         console.log(output);
       } else {