+++ src/fast-path.js
@@ -194,11 +194,24 @@ FastPath.prototype.needsParens = function() {
   }
 
   switch (node.type) {
-    case "CallExpression":
-      if (parent.type === "NewExpression" && parent.callee === node) {
+    case "CallExpression": {
+      let firstParentNotMemberExpression = parent;
+      let i = 0;
+      while (
+        firstParentNotMemberExpression &&
+        firstParentNotMemberExpression.type === "MemberExpression"
+      ) {
+        firstParentNotMemberExpression = this.getParentNode(++i);
+      }
+
+      if (
+        firstParentNotMemberExpression.type === "NewExpression" &&
+        firstParentNotMemberExpression.callee === this.getParentNode(i - 1)
+      ) {
         return true;
       }
       return false;
+    }
 
     case "SpreadElement":
     case "SpreadProperty":
@@ -405,8 +418,23 @@ FastPath.prototype.needsParens = function() {
         parent.type === "IntersectionTypeAnnotation"
       );
 
+    case "StringLiteral":
     case "NumericLiteral":
     case "Literal":
+      if (
+        typeof node.value === "string" &&
+        parent.type === "ExpressionStatement" &&
+        !parent.directive
+      ) {
+        // To avoid becoming a directive
+        const grandParent = this.getParentNode(1);
+
+        return (
+          grandParent.type === "Program" ||
+          grandParent.type === "BlockStatement"
+        );
+      }
+
       return (
         parent.type === "MemberExpression" &&
         typeof node.value === "number" &&
@@ -523,9 +551,6 @@ FastPath.prototype.needsParens = function() {
 
     case "ClassExpression":
       return parent.type === "ExportDefaultDeclaration";
-
-    case "StringLiteral":
-      return parent.type === "ExpressionStatement"; // To avoid becoming a directive
   }
 
   return false;